Выносит список городов в отдельный компонент CitiesList
